Projet_JEDHA_Kayak/requirements.txt
requests
json
pandas
datetime
numpy
plotly.express
plotly.graph_objects
plotly.io
matplotlib.pyplot
os
logging
scrapy
time
boto3
fsspec
s3fs